Total Redesign
Utilizing as much as possible from rift, all work done based on current
main branch.

New config setting:
[settings.layout.scroll]
# Maximum number of columns displayed at once in the scroll layout (1-5)
# This controls how many columns are visible simultaneously; extra
columns stay in the carousel.
visible_columns = 2

ALT+SHIFT+H/L (Still not perfect as I do plan on adding a command that
automatically takes out a focused window from a stacked column in a
newly created column to the right)

You can insert or remove a window/row from or to a column, max 3
windows/rows per column. I don't see the reason for beign mroe but if
there is a need then we can also add that setting to the config for
customization by the users.